change names of all JDBC directory
