+++ src/main/java/com/solvd/sql/jdbc/CategoryDao.java
@@ -1,7 +1,7 @@
 package com.solvd.sql.jdbc;
 
-import com.solvd.sql.interfaces.IDaoShop;
-import com.solvd.sql.model.Shop;
+import com.solvd.sql.interfaces.IDaoCategory;
+import com.solvd.sql.model.Category;
 import com.solvd.util.ConnectionPool;
 
 import java.sql.Connection;
@@ -11,19 +11,16 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class DaoShop implements IDaoShop {
+public class CategoryDao implements IDaoCategory {
     private final ConnectionPool connectionPool = ConnectionPool.getInstance();
 
     @Override
-    public void insert(Shop shop) throws SQLException {
+    public void insert(Category category) throws SQLException {
         Connection con = connectionPool.getConnection();
-        String query = "INSERT INTO shop (shop_name, address, phone, owner_id) VALUES (?, ?, ?, ?)";
+        String query = "INSERT INTO category (category_name) VALUES (?)";
 
         try (PreparedStatement ps = con.prepareStatement(query)) {
-            ps.setString(1, shop.getShopName());
-            ps.setString(2, shop.getAddress());
-            ps.setString(3, shop.getPhone());
-            ps.setInt(4, shop.getOwnerId());
+            ps.setString(1, category.getCategoryName());
             ps.execute();
         } catch (SQLException e) {
             throw new RuntimeException(e);
@@ -35,16 +32,13 @@ public class DaoShop implements IDaoShop {
     }
 
     @Override
-    public void update(Shop shop) throws SQLException {
+    public void update(Category category) throws SQLException {
         Connection con = connectionPool.getConnection();
-        String query = "UPDATE shop SET shop_name = ?, address = ?, phone = ?, owner_id = ? WHERE id = ?";
+        String query = "UPDATE category SET category_name = ? WHERE id = ?";
 
         try (PreparedStatement ps = con.prepareStatement(query)) {
-            ps.setString(1, shop.getShopName());
-            ps.setString(2, shop.getAddress());
-            ps.setString(3, shop.getPhone());
-            ps.setInt(4, shop.getOwnerId());
-            ps.setInt(5, shop.getId());
+            ps.setString(1, category.getCategoryName());
+            ps.setInt(2, category.getId());
             ps.execute();
         } catch (SQLException e) {
             throw new RuntimeException();
@@ -56,7 +50,7 @@ public class DaoShop implements IDaoShop {
     @Override
     public void delete(int id) throws SQLException {
         Connection con = connectionPool.getConnection();
-        String query = "DELETE FROM shop WHERE id = ?";
+        String query = "DELETE FROM category WHERE id = ?";
         try (PreparedStatement ps = con.prepareStatement(query)) {
             ps.setInt(1, id);
             ps.execute();
@@ -68,21 +62,18 @@ public class DaoShop implements IDaoShop {
     }
 
     @Override
-    public List<Shop> getAll() throws SQLException {
+    public List<Category> getAll() throws SQLException {
         Connection con = connectionPool.getConnection();
-        List<Shop> shops = new ArrayList<>();
-        String query = "SELECT * FROM shop";
+        List<Category> categories = new ArrayList<>();
+        String query = "SELECT * FROM category";
         try (PreparedStatement ps = con.prepareStatement(query)) {
             ps.execute();
             try (ResultSet rs = ps.getResultSet()) {
                 while (rs.next()) {
-                    Shop shop = new Shop();
-                    shop.setId(rs.getInt("id"));
-                    shop.setShopName(rs.getString("shop_name"));
-                    shop.setAddress(rs.getString("address"));
-                    shop.setPhone(rs.getString("phone"));
-                    shop.setOwnerId(rs.getInt("owner_id"));
-                    shops.add(shop);
+                    Category category = new Category();
+                    category.setId(rs.getInt("id"));
+                    category.setCategoryName(rs.getString("category_name"));
+                    categories.add(category);
                 }
             }
         } catch (SQLException e) {
@@ -90,24 +81,21 @@ public class DaoShop implements IDaoShop {
         } finally {
             connectionPool.releaseConnection(con);
         }
-        return shops;
+        return categories;
     }
 
     @Override
-    public Shop get(int id) throws SQLException {
+    public Category get(int id) throws SQLException {
         Connection con = connectionPool.getConnection();
-        Shop shop = new Shop();
-        String query = "SELECT * FROM shop WHERE id = ?";
+        Category category = new Category();
+        String query = "SELECT * FROM category WHERE id = ?";
         try (PreparedStatement ps = con.prepareStatement(query)) {
             ps.setInt(1, id);
-
             ps.execute();
             try (ResultSet rs = ps.getResultSet()) {
                 while (rs.next()) {
-                    shop.setShopName(rs.getString("shop_name"));
-                    shop.setAddress(rs.getString("address"));
-                    shop.setPhone(rs.getString("phone"));
-                    shop.setOwnerId(rs.getInt("owner_id"));
+                    category.setId(rs.getInt("id"));
+                    category.setCategoryName(rs.getString("category_name"));
                 }
             } catch (SQLException e) {
                 throw new RuntimeException();
@@ -115,6 +103,28 @@ public class DaoShop implements IDaoShop {
                 connectionPool.releaseConnection(con);
             }
         }
-        return shop;
+        return category;
+    }
+
+    @Override
+    public Category get(String categoryName) throws SQLException {
+        Connection con = connectionPool.getConnection();
+        Category category = new Category();
+        String query = "SELECT * FROM category WHERE category_name = ?";
+        try (PreparedStatement ps = con.prepareStatement(query)) {
+            ps.setString(1, categoryName);
+            ps.execute();
+            try (ResultSet rs = ps.getResultSet()) {
+                while (rs.next()) {
+                    category.setId(rs.getInt("id"));
+                    category.setCategoryName(rs.getString("category_name"));
+                }
+            } catch (SQLException e) {
+                throw new RuntimeException();
+            } finally {
+                connectionPool.releaseConnection(con);
+            }
+        }
+        return category;
     }
 }
