+++ src/main/java/com/solvd/sql/jdbc/CustomerDao.java
@@ -1,7 +1,7 @@
 package com.solvd.sql.jdbc;
 
-import com.solvd.sql.interfaces.IDaoSupplier;
-import com.solvd.sql.model.Supplier;
+import com.solvd.sql.interfaces.IDaoCustomer;
+import com.solvd.sql.model.Customer;
 import com.solvd.util.ConnectionPool;
 
 import java.sql.Connection;
@@ -11,18 +11,16 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class DaoSupplier implements IDaoSupplier {
+public class CustomerDao implements IDaoCustomer {
     private final ConnectionPool connectionPool = ConnectionPool.getInstance();
-
     @Override
-    public void insert(Supplier supplier) throws SQLException {
+    public void insert(Customer customer) throws SQLException {
         Connection con = connectionPool.getConnection();
-        String query = "INSERT INTO supplier (supplier_name, tax_number, phone) VALUES (?, ?, ?)";
+        String query = "INSERT INTO customer (tax_number, person_id) VALUES (?, ?)";
 
         try (PreparedStatement ps = con.prepareStatement(query)) {
-            ps.setString(1, supplier.getSupplierName());
-            ps.setString(2, supplier.getTaxNumber());
-            ps.setString(3, supplier.getPhone());
+            ps.setString(1, customer.getTaxNumber());
+            ps.setInt(1, customer.getPersonId());
             ps.execute();
         } catch (SQLException e) {
             throw new RuntimeException(e);
@@ -34,15 +32,14 @@ public class DaoSupplier implements IDaoSupplier {
     }
 
     @Override
-    public void update(Supplier supplier) throws SQLException {
+    public void update(Customer customer) throws SQLException {
         Connection con = connectionPool.getConnection();
-        String query = "UPDATE supplier SET supplier_name = ?, tax_number = ?, phone = ? WHERE id = ?";
+        String query = "UPDATE customer SET tax_number = ?, person_id = ?, WHERE id = ?";
 
         try (PreparedStatement ps = con.prepareStatement(query)) {
-            ps.setString(1, supplier.getSupplierName());
-            ps.setString(2, supplier.getTaxNumber());
-            ps.setString(3, supplier.getPhone());
-            ps.setInt(5, supplier.getId());
+            ps.setString(1, customer.getTaxNumber());
+            ps.setInt(2, customer.getPersonId());
+            ps.setInt(3, customer.getId());
             ps.execute();
         } catch (SQLException e) {
             throw new RuntimeException();
@@ -54,7 +51,7 @@ public class DaoSupplier implements IDaoSupplier {
     @Override
     public void delete(int id) throws SQLException {
         Connection con = connectionPool.getConnection();
-        String query = "DELETE FROM supplier WHERE id = ?";
+        String query = "DELETE FROM customer WHERE id = ?";
         try (PreparedStatement ps = con.prepareStatement(query)) {
             ps.setInt(1, id);
             ps.execute();
@@ -66,20 +63,19 @@ public class DaoSupplier implements IDaoSupplier {
     }
 
     @Override
-    public List<Supplier> getAll() throws SQLException {
+    public List<Customer> getAll() throws SQLException {
         Connection con = connectionPool.getConnection();
-        List<Supplier> suppliers = new ArrayList<>();
-        String query = "SELECT * FROM supplier";
-        try (PreparedStatement ps = con.prepareStatement(query)) {
+        List<Customer> customers = new ArrayList<>();
+        String query = "SELECT * FROM customer";
+        try(PreparedStatement ps = con.prepareStatement(query)) {
             ps.execute();
             try (ResultSet rs = ps.getResultSet()) {
-                while (rs.next()) {
-                    Supplier supplier = new Supplier();
-                    supplier.setId(rs.getInt("id"));
-                    supplier.setSupplierName(rs.getString("supplier_name"));
-                    supplier.setTaxNumber(rs.getString("tax_number"));
-                    supplier.setPhone(rs.getString("phone"));
-                    suppliers.add(supplier);
+                while ( rs.next()) {
+                    Customer customer = new Customer();
+                    customer.setId(rs.getInt("id"));
+                    customer.setTaxNumber(rs.getString("tax_number"));
+                    customer.setPersonId(rs.getInt("person_id"));
+                    customers.add(customer);
                 }
             }
         } catch (SQLException e) {
@@ -87,24 +83,22 @@ public class DaoSupplier implements IDaoSupplier {
         } finally {
             connectionPool.releaseConnection(con);
         }
-        return suppliers;
+        return customers;
     }
 
     @Override
-    public Supplier get(int id) throws SQLException {
+    public Customer get(int id) throws SQLException {
         Connection con = connectionPool.getConnection();
-        Supplier supplier = new Supplier();
-        String query = "SELECT * FROM supplier WHERE id = ?";
+        Customer customer = new Customer();
+        String query = "SELECT * FROM customer WHERE id = ?";
         try (PreparedStatement ps = con.prepareStatement(query)) {
             ps.setInt(1, id);
-
             ps.execute();
             try (ResultSet rs = ps.getResultSet()) {
                 while (rs.next()) {
-                    supplier.setId(rs.getInt("id"));
-                    supplier.setSupplierName(rs.getString("supplier_name"));
-                    supplier.setTaxNumber(rs.getString("tax_number"));
-                    supplier.setPhone(rs.getString("phone"));
+                    customer.setId(rs.getInt("id"));
+                    customer.setTaxNumber(rs.getString("tax_number"));
+                    customer.setPersonId(rs.getInt("person_id"));
                 }
             } catch (SQLException e) {
                 throw new RuntimeException();
@@ -112,6 +106,6 @@ public class DaoSupplier implements IDaoSupplier {
                 connectionPool.releaseConnection(con);
             }
         }
-        return supplier;
+        return customer;
     }
 }
