+++ src/main/java/com/solvd/sql/jdbc/ProductDao.java
@@ -1,7 +1,7 @@
 package com.solvd.sql.jdbc;
 
-import com.solvd.sql.interfaces.IDaoPerson;
-import com.solvd.sql.model.Person;
+import com.solvd.sql.interfaces.IDaoProduct;
+import com.solvd.sql.model.Product;
 import com.solvd.util.ConnectionPool;
 
 import java.sql.Connection;
@@ -11,19 +11,20 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class DaoPerson implements IDaoPerson {
+public class ProductDao implements IDaoProduct {
     private final ConnectionPool connectionPool = ConnectionPool.getInstance();
 
     @Override
-    public void insert(Person person) throws SQLException {
+    public void insert(Product product) throws SQLException {
         Connection con = connectionPool.getConnection();
-        String query = "INSERT INTO person (person_name, last_name, phone, address) VALUES (?, ?, ?, ?)";
+        String query = "INSERT INTO product (product_name, stock, price, category_id, supplier_id) VALUES (?, ?, ?, ?, ?)";
 
         try (PreparedStatement ps = con.prepareStatement(query)) {
-            ps.setString(1, person.getPersonName());
-            ps.setString(2, person.getLastName());
-            ps.setString(3, person.getPhone());
-            ps.setString(4, person.getAddress());
+            ps.setString(1, product.getProductName());
+            ps.setInt(2, product.getStock());
+            ps.setDouble(3, product.getPrice());
+            ps.setInt(4, product.getCategoryId());
+            ps.setInt(5, product.getSupplierId());
             ps.execute();
         } catch (SQLException e) {
             throw new RuntimeException(e);
@@ -35,16 +36,17 @@ public class DaoPerson implements IDaoPerson {
     }
 
     @Override
-    public void update(Person person) throws SQLException {
+    public void update(Product product) throws SQLException {
         Connection con = connectionPool.getConnection();
-        String query = "UPDATE person SET person_name = ?, last_name = ?, phone = ?, address = ? WHERE id = ?";
+        String query = "UPDATE product SET product_name = ?, stock = ?, price = ?, category_id = ?, supplier_id = ? WHERE id = ?";
 
         try (PreparedStatement ps = con.prepareStatement(query)) {
-            ps.setString(1, person.getPersonName());
-            ps.setString(2, person.getLastName());
-            ps.setString(3, person.getPhone());
-            ps.setString(4, person.getAddress());
-            ps.setInt(5, person.getId());
+            ps.setString(1, product.getProductName());
+            ps.setInt(2, product.getStock());
+            ps.setDouble(3, product.getPrice());
+            ps.setInt(4, product.getCategoryId());
+            ps.setInt(5, product.getSupplierId());
+            ps.setInt(6, product.getId());
             ps.execute();
         } catch (SQLException e) {
             throw new RuntimeException();
@@ -56,7 +58,7 @@ public class DaoPerson implements IDaoPerson {
     @Override
     public void delete(int id) throws SQLException {
         Connection con = connectionPool.getConnection();
-        String query = "DELETE FROM person WHERE id = ?";
+        String query = "DELETE FROM product WHERE id = ?";
         try (PreparedStatement ps = con.prepareStatement(query)) {
             ps.setInt(1, id);
             ps.execute();
@@ -68,21 +70,22 @@ public class DaoPerson implements IDaoPerson {
     }
 
     @Override
-    public List<Person> getAll() throws SQLException {
+    public List<Product> getAll() throws SQLException {
         Connection con = connectionPool.getConnection();
-        List<Person> persons = new ArrayList<>();
-        String query = "SELECT * FROM person";
+        List<Product> products = new ArrayList<>();
+        String query = "SELECT * FROM product";
         try (PreparedStatement ps = con.prepareStatement(query)) {
             ps.execute();
             try (ResultSet rs = ps.getResultSet()) {
                 while (rs.next()) {
-                    Person person = new Person();
-                    person.setId(rs.getInt("id"));
-                    person.setPersonName(rs.getString("person_name"));
-                    person.setLastName(rs.getString("last_name"));
-                    person.setPhone(rs.getString("phone"));
-                    person.setAddress(rs.getString("address"));
-                    persons.add(person);
+                    Product product = new Product();
+                    product.setId(rs.getInt("id"));
+                    product.setProductName(rs.getString("product_name"));
+                    product.setStock(rs.getInt("stock"));
+                    product.setPrice(rs.getDouble("price"));
+                    product.setCategoryId(rs.getInt("category_id"));
+                    product.setSupplierId(rs.getInt("supplier_id"));
+                    products.add(product);
                 }
             }
         } catch (SQLException e) {
@@ -90,25 +93,26 @@ public class DaoPerson implements IDaoPerson {
         } finally {
             connectionPool.releaseConnection(con);
         }
-        return persons;
+        return products;
     }
 
     @Override
-    public Person get(int id) throws SQLException {
+    public Product get(int id) throws SQLException {
         Connection con = connectionPool.getConnection();
-        Person person = new Person();
-        String query = "SELECT * FROM person WHERE id = ?";
+        Product product = new Product();
+        String query = "SELECT * FROM product WHERE id = ?";
         try (PreparedStatement ps = con.prepareStatement(query)) {
             ps.setInt(1, id);
 
             ps.execute();
             try (ResultSet rs = ps.getResultSet()) {
                 while (rs.next()) {
-                    person.setId(rs.getInt("id"));
-                    person.setPersonName(rs.getString("person_name"));
-                    person.setLastName(rs.getString("last_name"));
-                    person.setPhone(rs.getString("phone"));
-                    person.setAddress(rs.getString("address"));
+                    product.setId(rs.getInt("id"));
+                    product.setProductName(rs.getString("product_name"));
+                    product.setStock(rs.getInt("stock"));
+                    product.setPrice(rs.getDouble("price"));
+                    product.setCategoryId(rs.getInt("category_id"));
+                    product.setSupplierId(rs.getInt("supplier_id"));
                 }
             } catch (SQLException e) {
                 throw new RuntimeException();
@@ -116,6 +120,6 @@ public class DaoPerson implements IDaoPerson {
                 connectionPool.releaseConnection(con);
             }
         }
-        return person;
+        return product;
     }
 }
